Improve 'location' display in list of reports
The list of reports should show location to best appropriate granularity, ie the
most suitable field below the field with common values. For a geographically
broad list, this might be cities; for a limited list, it might be street name.
Geocoding data returned by Google seems to be quite variable, so returning the
best value is a rather complex operation.