Implement partial submissions on page-by-page incident report submissions
This creates a stub incident report in the database as soon as a report
submission is started, which gets added to with each page submitted.

Rather complex code handles the case of checkboxes / radio-buttons which don't get
POSTed if they are not selected; ensuring these appear (as null) in the submitted
report (in the correct order), whilst fields from pages not visited in an
incomplete submbission are not created at all, was complex, and can perhaps be
improved on.