Rationalise postPage logic, refactor & clean up code
In the process of enabling partial report submissions, then moving the
'used-before' page from the end to the beginning, the postPage() logic became
horribly convoluted. With the decision not to be able to go back from p2 to p1
(used-before), the logic can be rationalised massively. Also, inconsistencies &
errors in the flash notifications were corrected, and the show/hide of
'used-before' options fixed.