Fix tests for new session, to correct 'expired session' & 'jump ahead' checks
Setting session to null destroys the session, but the next request sees an empty
object, so !ctx.session is not a good test: this uses session.isNew instead. This
means requesting p1 before visiting the index page causes a redirect, and fixes
'your session has expired' behaviour. Note that both of these cases are hard to
write tests for.