Make 'button' a pure button rather than a submit button
The 'button' element is currently used only for generating a new alias, which
should not submit the form. If a 'submit' button should actually be required,
we can create a new 'submit' element in the yaml for it, but at the moment the
only submit button is the automatically-generated 'submit and continue' button.